sensor-agent/requirements.txt
# Core Components
requests>=2.31.0
watchdog>=4.0.0
python-dotenv>=1.0.1
pydantic>=2.6.3

# Networking
httpx>=0.27.0

# Development
pytest>=8.1.1
black>=24.2.0
